make gcc happy

+++ src/tsutil.h
@@ -14,12 +14,13 @@
 #define ALDO_NS_PER_MS 1e6
 
 #include "bridgeopen.h"
-aldo_const int AldoNsPerS = (int)(ALDO_MS_PER_S * ALDO_NS_PER_MS);
+aldo_const int AldoNsPerS = (int)ALDO_MS_PER_S * (int)ALDO_NS_PER_MS;
 
 aldo_export
 inline double aldo_timespec_to_ms(const struct timespec *ts) aldo_nothrow
 {
-    return (ts->tv_sec * ALDO_MS_PER_S) + (ts->tv_nsec / ALDO_NS_PER_MS);
+    return ((double)ts->tv_sec * ALDO_MS_PER_S)
+            + ((double)ts->tv_nsec / ALDO_NS_PER_MS);
 }
 
 aldo_export
